+++ src/Console/ExitStatusListener.php
@@ -5,7 +5,7 @@ declare(strict_types = 1);
 namespace hanneskod\readmetester\Console;
 
 use hanneskod\readmetester\ListenerInterface;
-use hanneskod\readmetester\Example\Example;
+use hanneskod\readmetester\Example\ExampleInterface;
 use hanneskod\readmetester\Expectation\Status;
 
 /**
@@ -18,11 +18,11 @@ class ExitStatusListener implements ListenerInterface
      */
     private $exitStatus = 0;
 
-    public function onExample(Example $example): void
+    public function onExample(ExampleInterface $example): void
     {
     }
 
-    public function onIgnoredExample(Example $example): void
+    public function onIgnoredExample(ExampleInterface $example): void
     {
     }