+++ src/Example/Example.php
@@ -5,41 +5,31 @@ declare(strict_types = 1);
 namespace hanneskod\readmetester\Example;
 
 use hanneskod\readmetester\Expectation\ExpectationInterface;
+use hanneskod\readmetester\Name\NameInterface;
 use hanneskod\readmetester\Parser\CodeBlock;
 
-/**
- * Default example
- */
-class Example
+class Example implements ExampleInterface
 {
-    /**
-     * @var string
-     */
+    /** @var NameInterface */
     private $name;
 
-    /**
-     * @var CodeBlock
-     */
+    /** @var CodeBlock */
     private $code;
 
-    /**
-     * @var ExpectationInterface[]
-     */
+    /** @var ExpectationInterface[] */
     private $expectations;
 
     /**
-     * @param string                 $name         Name of this example
-     * @param CodeBlock              $code         Example code
-     * @param ExpectationInterface[] $expectations List of expectations
+     * @param ExpectationInterface[] $expectations
      */
-    public function __construct(string $name, CodeBlock $code, array $expectations)
+    public function __construct(NameInterface $name, CodeBlock $code, array $expectations)
     {
         $this->name = $name;
         $this->code = $code;
         $this->expectations = $expectations;
     }
 
-    public function getName(): string
+    public function getName(): NameInterface
     {
         return $this->name;
     }