+++ src/Example/ExampleInterface.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace hanneskod\readmetester\Example;
+
+use hanneskod\readmetester\Expectation\ExpectationInterface;
+use hanneskod\readmetester\Name\NameInterface;
+use hanneskod\readmetester\Parser\CodeBlock;
+
+interface ExampleInterface
+{
+    /**
+     * Get example name
+     */
+    public function getName(): NameInterface;
+
+    /**
+     * Get example code block
+     */
+    public function getCodeBlock(): CodeBlock;
+
+    /**
+     * Get expectations associated with example
+     *
+     * @return ExpectationInterface[]
+     */
+    public function getExpectations(): array;
+}