+++ src/Name/AnonymousName.php
@@ -4,17 +4,13 @@ declare(strict_types = 1);
 
 namespace hanneskod\readmetester\Name;
 
-final class ExampleName implements NameInterface
+final class AnonymousName implements NameInterface
 {
-    /** @var string */
-    private $shortName;
-
     /** @var string */
     private $namespace;
 
-    public function __construct(string $shortName, string $namespace)
+    public function __construct(string $namespace)
     {
-        $this->shortName = $shortName;
         $this->namespace = $namespace;
     }
 
@@ -25,7 +21,7 @@ final class ExampleName implements NameInterface
 
     public function getShortName(): string
     {
-        return $this->shortName;
+        return 'UNNAMED';
     }
 
     public function getNamespaceName(): string
@@ -35,6 +31,6 @@ final class ExampleName implements NameInterface
 
     public function equals(NameInterface $name): bool
     {
-        return $name->getCompleteName() == $this->getCompleteName();
+        return false;
     }
 }