+++ tests/Example/ExampleTest.php
@@ -5,22 +5,25 @@ declare(strict_types = 1);
 namespace hanneskod\readmetester\Example;
 
 use hanneskod\readmetester\Expectation\ExpectationInterface;
+use hanneskod\readmetester\Name\NameInterface;
 use hanneskod\readmetester\Parser\CodeBlock;
 
 class ExampleTest extends \PHPUnit\Framework\TestCase
 {
     function testGetName()
     {
+        $name = $this->createMock(NameInterface::CLASS);
         $this->assertSame(
-            'foobar',
-            (new Example('foobar', $this->prophesize(CodeBlock::CLASS)->reveal(), []))->getName()
+            $name,
+            (new Example($name, $this->createMock(CodeBlock::CLASS), []))->getName()
         );
     }
 
     function testShouldBeEaluated()
     {
         $this->assertTrue(
-            (new Example('', $this->prophesize(CodeBlock::CLASS)->reveal(), []))->shouldBeEvaluated()
+            (new Example($this->createMock(NameInterface::CLASS), $this->createMock(CodeBlock::CLASS), []))
+                ->shouldBeEvaluated()
         );
     }
 
@@ -29,16 +32,17 @@ class ExampleTest extends \PHPUnit\Framework\TestCase
         $code = $this->prophesize(CodeBlock::CLASS)->reveal();
         $this->assertSame(
             $code,
-            (new Example('', $code, []))->getCodeBlock()
+            (new Example($this->createMock(NameInterface::CLASS), $code, []))->getCodeBlock()
         );
     }
 
     function testGetExpectations()
     {
-        $expectations = [$this->prophesize(ExpectationInterface::CLASS)->reveal()];
+        $expectations = [$this->createMock(ExpectationInterface::CLASS)];
         $this->assertSame(
             $expectations,
-            (new Example('', $this->prophesize(CodeBlock::CLASS)->reveal(), $expectations))->getExpectations()
+            (new Example($this->createMock(NameInterface::CLASS), $this->createMock(CodeBlock::CLASS), $expectations))
+                ->getExpectations()
         );
     }
 }