+++ tests/Example/IgnoredExampleTest.php
@@ -4,6 +4,7 @@ declare(strict_types = 1);
 
 namespace hanneskod\readmetester\Example;
 
+use hanneskod\readmetester\Name\NameInterface;
 use hanneskod\readmetester\Parser\CodeBlock;
 
 class IgnoredExampleTest extends \PHPUnit\Framework\TestCase
@@ -11,7 +12,8 @@ class IgnoredExampleTest extends \PHPUnit\Framework\TestCase
     function testShouldBeEaluated()
     {
         $this->assertFalse(
-            (new IgnoredExample('', $this->prophesize(CodeBlock::CLASS)->reveal(), []))->shouldBeEvaluated()
+            (new IgnoredExample($this->createMock(NameInterface::CLASS), $this->createMock(CodeBlock::CLASS), []))
+                ->shouldBeEvaluated()
         );
     }
 }