+++ CHANGES.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Renamed `--flagged-only` => `--named-only`
+
 # Changes in 1.0-beta4
 
 * `@include` annotation replaces `@extends`.